Made the curves and lines smoother!
Also, added a new line, which shows that they aren't *quite* as smooth
as they oughta be.
In the end I just made points stored with floating point coordinates,
which is converted to integral stuff at the last possible moment before
drawing.

+++ src/point_and_color.rs
@@ -2,7 +2,7 @@ use std::ops::{Add, Mul};
 use crate::PixelGrid;
 
 #[derive(Clone, Copy)]
-pub struct Point2D(pub u64, pub u64); // A point in 2d space, represented as (x, y), with (0, 0) as the upper-left corner.
+pub struct Point2D(pub f64, pub f64); // A point in 2d space, represented as (x, y), with (0, 0) as the upper-left corner.
 #[derive(Clone, Copy)]
 pub struct Color(pub u8, pub u8, pub u8); // An 8-bit-per-channel/24-bit-total color, suitable for "millions" (about 16.7 million) colors.
 
@@ -10,7 +10,7 @@ pub struct Color(pub u8, pub u8, pub u8); // An 8-bit-per-channel/24-bit-total c
 impl Mul<f64> for Point2D {
     type Output = Point2D;
     fn mul(self, rhs: f64) -> Self::Output {
-        Point2D((self.0 as f64 * rhs) as u64, (self.1 as f64 * rhs) as u64)
+        Point2D(self.0 * rhs, self.1 * rhs)
     }
 }
